refactor tests to remove unused test dependencies
