+++ example/graphqldev/go.mod
@@ -9,7 +9,6 @@ require (
 
 require (
 	github.com/google/uuid v1.6.0 // indirect
-	github.com/graph-gophers/graphql-transport-ws v0.0.2 // indirect
 	nhooyr.io/websocket v1.8.10 // indirect
 )
